ovn-northd-ddlog: Avoid N*M crossproduct joining switches with routers.
DDlog takes a literalist view of joins, executing them in the order that
are written without attempting much in the way of reordering or
optimization.  The FirstHopLogicalRouter rule as written here joined
LogicalRouterPort with LogicalSwitchPort without using any join key,
and then later eliminated some of the possibilities.  This meant that
if there were N router ports and M switch ports, DDlog actually
considered all N*M possibilities, which is expensive.

This commit improves the big-O of the situation by introducing an
intermediate table that contains only the switch port that connect to
router ports and by giving that table a column that can be used as a
join key.  This allows DDlog to join them efficiently.  (The new column
is needed because DDlog cannot join on a member of a map directly.)

Found via the DDlog profiling feature:
https://github.com/vmware/differential-datalog/blob/master/doc/tutorial/tutorial.md#Profiling

Suggested-by: Mihai Budiu <mbudiu@vmware.com>
Suggested-by: Leonid Ryhzyk <lryzhyk@vmware.com>
Signed-off-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>