tests: Make certificate generation extendable
In preparation for enabling testing with SSL and RBAC enabled by
default, rework the certificate generation so that we can easily
add generation of more certificates/CN on demand.

A side erffect of the change is a more generic naming scheme for
the certificate files so the patch also contains an update to
existing tests so that they use the new filenames.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>