ovn-northd.c: Don't use DPG for MC_UNKNOWN related lflows.
In commit b468c2c1 it avoided using DPG for multicast related lflows, but
commit dd94f126 update the MC_UNKNOWN related lflows with DPG used again
by mistake. This patch fixes it to avoid problems when a lflow using
MC_UNKNOWN is monitored by a ovn-controller before the related mc-group
is monitored, which could cause ovn-controller problem due to an
incomplete dependency handling in I-P.

This change can be removed (and applying DPG for all the other mc-group
related lflows) when the I-P dependency handling problem is fixed.

This change also fixes the ovn-northd-ddlog part, which missed using
UniqueFlow() for MC_UNKNOWN from the initial ddlog commit 0e77b3bcbf.

Fixes: dd94f1266c ("northd: MAC learning: Add logical flows for fdb.")
Fixes: 0e77b3bcbf ("ovn-northd-ddlog: New implementation of ovn-northd based on ddlog.")
Cc: Numan Siddique <numans@ovn.org>
Cc: Leonid Ryzhyk <lryzhyk@vmware.com>
Co-authored-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>