physical: Set the port binding uuid as cookie for flows where relevant.
Some of the OF flows added by physical.c in tables LOCAL_OUTPUT,
CHECK_LOOPBACK, SAVE_INPORT can store the port binding uuid as
flow cookie because these flows matches on port binding key
in the MFF_LOG_OUTPORT field.  Whereas right now it stores 0 as cookie.

This patch stores the port binding uuid as cookie for these flows.
When a port binding is deleted, the port binding handler for flow_output
engine node flood removes the flows with the port binding uuid and
such flows wont get removed.  Right now this is not a problem because
deleting a port binding also triggers in physical_run() being called
and it recomputes all the physical flows.

This patch is required for an upcoming patch which do not call
physical_run() for port binding deletes.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>