ovn-controller: Fix port-group incremental processing.
When SB port-group (per DP) is deleted and added back and the
notification to ovn-controller include both operations in a reversed
order, the current change handler in ovn-controller would end up
deleting the SB port-group in the local cache, which in turn result in
missing flows, and even worse it can result in crash if runtime data
handler is triggered later because we asserts the SB PG should be
equivalent to the local cache.

This patch fixes it by always handling deletion for tracked PG changes,
just like how we are handling other tracked changes in I-P (e.g.
logical_flow). A test case is crafted to cover such scenario.

Fixes: 0cfeba6b55 ("ovn-controller: Fix port group conjunction flow explosion problem.")
Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>