ovn.at: Fix test "No ovn-controller assert when generating conjunction flows -- ovn-northd-ddlog".
The test fails quite often for ddlog because of tunnel keys mismatch when
checking OpenFlow dumps. It is possible to allocate different keys in
different runs because multiple NB changes can be handled by northd in a
single iteration - it fails more often in ddlog probably because there
is some more latency for northd-ddlog to start processing the changes,
but it doesn't affect correctness. So this patch fixes it by ignoring
the tunnel keys when comparing the flows.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>