lflow-cache: Auto trim when cache size is reduced significantly.
Due to a bug in glibc [0], M_TRIM_THRESHOLD (default 128KB) is not
honored.  The lflow cache is one of the largest memory consumers in
ovn-controller and it used to trim memory whenever the cache was
flushed.  However, that required periodic intervention from the CMS
side.

Instead, we now automatically trim memory every time the lflow cache
utilization decreases by 50%, with a threshold of at least
lflow-cache-trim-limit (10000) elements in the cache.

The percentage of the high watermark under which memory is trimmed
automatically as well as the lflow-cache minimum number of elements
above which trimming is performed are configurable through OVS external
IDs.

[0] https://sourceware.org/bugzilla/show_bug.cgi?id=14827

Reported-at: https://bugzilla.redhat.com/1967882
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>