tests: Fix "vlan traffic for external network with distributed..."
This fixes the test case "ovn -- vlan traffic for external network with
distributed router gateway port", which often failed, especially under
ddlog.  The problem was that the test case was setting a column that
only ovn-northd is supposed to set.  The "tag" column in
Logical_Switch_Port is for ovn-northd to report the tag for a logical
switch port.  It's documented that way in ovn-nb(5):

       tag: optional integer, in range 1 to 4,095
              The  VLAN tag allocated by ovn-northd based on the contents of the
              tag_request column.

The right way to request a particular tag is to use tag_request.  This
commit fixes that.

This also fixes the similar problem I found in the test "ovn -- send
gratuitous ARP for NAT rules on HA distributed router", even though I
hadn't seen similar failures there.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>