ovn-northd-ddlog: Avoid N**2 blowup for N connected logical routers.
It's easy to implement "connected components" in raw DDlog, but it
takes N**2 time and space in the number of elements in a component.
This was a huge waste for a test case supplied by Dumitru Ceara that
had over 8000 logical routers.  This commit solves the problem by using
the "graph" transformer built in DDlog, which efficiently implements
connected components.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Reported-by: Dumitru Ceara <dceara@redhat.com>
Reported-at: https://mail.openvswitch.org/pipermail/ovs-dev/2021-June/384519.html
Signed-off-by: 0-day Robot <robot@bytheb.org>