binding: Add the localport port binding in the binding_lport information.
If there is an OVS interface present with the external_ids:iface-id set
to a localport port binding, we create a 'struct local_binding' for this
OVS interface but we do not associate the binding_lport to this
local_binding.  This patch now associates the binding_lport now.
Future patch will make use of this information in handling the
runtime data changes in the pflow engine node.

Acked-by: Han Zhou <hzhou@ovn.org>
Reviewed-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>