I-P: Handle runtime data changes for pflow_output engine.
physical_run() maintains a local copy of local vif to ofports
in a simap along with the chassis tunnel information.  This patch
removes this from the physical module and now stores it in a
separate engine node - non_vif_data.  This makes it easier to
handle runtime data changes in pflow_output engine.

The newly added handler pflow_output_runtime_data_handler() returns
false if a datapath is added or removed from the local_datapaths
and handles the logical port claims and releases incrementally.

Acked-by: Han Zhou <hzhou@ovn.org>
Reviewed-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>