controller: Improve ct zone handling.
Prior to this patch, ovn-controller generates a zone id for each
OVS interface which has external_ids:iface-id set even if there is
no corresponding logical port for it.  This patch now changes the
way we allocate the zone id.  A zone id is allocated only if
there is an OVS interface with external_ids:iface-id and the
corresponding logical port is claimed.  We use the runtime data
(rt_data->lbinding_data.lports) for this.

This patch also improves the ct_zones_runtime_data_handler()
by using the tracked datapath data to allocate the zone ids
for newly claimed lports or to free up the zone id for
the released lports instead of triggering a full recompute.

And finally this patch also adds a ct zone handler for pflow_output
engine.  This handler falls back to recompute if the ct zone engine
was recomputed.  Otherwise it returns true.

Acked-by: Han Zhou <hzhou@ovn.org>
Reviewed-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>