northd: Alter initial hash sizing for dp_groups+parallel
When running with dp_groups and parallelization enabled we have
a possible worse case scenario where northd connects for the
first time to a pre-populated database.
If we do not size the hash to a sufficiently big size to
accommodate for this, we end up with severe lock contention
and very slow lookups.
If both dp_groups and parallelization are enabled we have
no choice, but to allocate for a worst case scenario on the
first run and adjust later.

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>