northd: Split northd.c
This commit splits northd into two parts. One part, ovn-northd.c,
is responsible for the `ovn-northd` application, and the other is
responsible for northd processing.

This takes one step towards a more modular northd code base.

It also allows northd to be used as a library that could be included,
for example, into the incremental processing framework.

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>