ovn-northd-ddlog: Reverse order of joins for connection tracking flows.
DDlog evaluates rules in the order given by syntax.  The rules for
load balancers all first evaluated a Router or Switch, then joined
that against the load balancers.  However, the expensive logic was
all on the load balancers.  This meant that the load balancer logic
was happening many times, once per switch or router that contained
it.  When a single load balancer was part of many switches or
routers, this did a lot of redundant processing.  This commit
reverses the join order, which speeds up processing a lot.

This commit looks big because it also converts a lot of rules from
the FLWOR syntax to traditional Datalog-style syntax.  This is not
completely needed, but the Datalog-style syntax is more versatile
(it supports FlatMap and aggregation), so I tend to make that sort
of change as I refactor things.

Signed-off-by: Ben Pfaff <blp@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>