binding: Store timestamp at which ovn-installed was set.
CMSs (e.g., ovn-kubernetes) already use the ovn-installed
external_id set by ovn-controller in OVS.Interface.external_ids to
determine that the networking for a VIF is completely plugged.

ovn-controller now also stores the timestamp (in milliseconds since the
epoch) at which an interface was marked as "installed", as the
ovn-installed-ts external id in the OVS DB.

This commit also adds the relevant documentation for
ovn-installed/ovn-installed-ts.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>