controller: configure only matching encaps between chassis
Previously tunnels encap on a chassis was based on the remote chassis
"best" encap type. Suppose we have 2 chassis: one with STT, another with
GENEVE. In this case on chassis 1 there was configured STT tunnel, and
GENEVE on another one. No traffic could be sent between these chassis.

With this approach it was impossible to change encap type
for chassis one-by-one, because different tunnel types were
configured on different edges of the link.
Suppose we have 2 chassis: one with STT and VXLAN configured encaps,
another with GENEVE and STT. In
this case on chassis 1 there was configured STT tunnel (best of VXLAN
and STT) and GENEVE on another one ("best" of GENEVE and STT).
No traffic could be sent between these chassis. Though the common STT
could be used.

Now we configure only matching encaps between nodes.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>