ovn.at: Fix flaky test "controller I-P handling with monitoring disabled".
Test was waiting for port to be up in SBDB before checking number of flows
in OVS. However, with conditional monitoring, there is no guarantee that all
flows are installed in OVS when port is up.
In addition, in both conditional monitoring and monitor-all cases, ARP
reponder flows are installed when port is detected up.
Hence, test was randomly failing as some flows were installed, but not all.
To fix this, we wait until the ARP related flows are installed.
Also fixed small typo in logging (for the same test).

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2004390
Fixes: f8a81693b0 ("ovn-controller: Fix the missing flows with monitor-all set to True")
Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>