controller: Allocate zone ids for localport port bindings.
Commit d4bca93 limited ct zone allocation to VIF-ports only, this
breaks OpenStack's use of localport for providing metadata to
instances.

This patch adds lports of type 'localport' to the list of lport
types to allocate zone ids for.

No test case is added as I am unable to make the error condition
occur reliably in the sandbox test environment, however I have
confirmed that the error condition is there on an unmodified OVN
installation and that the behavior changes by adding this patch.
Both for a fresh deployment where a modified package is slipped in
prior to first start of OVN and comparison of patched and
non-patched OVN after reboot of a deployed node.

Reported-at: https://mail.openvswitch.org/pipermail/ovs-discuss/2021-September/051473.html
Fixes: d4bca93 ("controller: Don't allocate zone ids for non-VIF port bindings. ")
Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>