libs: Replace weak with strong compare and exchange
Weak may fail for reasons unrelated to the comparison (not
on x86, where these are practically equivalent).

Thus, it can and should be used only in cases where there will
be a repeat of the event which runs the code (in this case
semaphore post).

On non-x86 architectures using weak may result in a failed
comparison without any more sem post events - a hang. Reported
for arm, likely for other platforms where weak may fail.

Reported-by: wentao.jia@easystack.cn

Signed-off-by: Anton Ivanov <anton.ivanov@cambridgegreys.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>