controller: Consider plugging VIF on CMS request.
When OVN is linked with an appropriate VIF plug provider,
CMS can request an OVN controller to plug individual lports into
the Open vSwitch instance it manages.

The port and interface record will be maintained throughout the
lifetime of the lport and it will be removed on release of lport.

Signed-off-by: Frode Nordahl <frode.nordahl@canonical.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>