vtep: set is-vtep to chassis's other_config if absent
After commit [0] the mandatory vtep chassis's option 'is-vtep' has appeared.
The upgrade scenario for ovn-controller-vtep was not supported:
'is-vtep' option was set only on vtep chassis creation. If chassis was
created prior to a new codebase, it was left intact and HW VTEP connectivity
was broken.
This commit fixes such scenario and now 'is-vtep' is set for vtep chassis
other_config if was not set yet.

0: https://github.com/ovn-org/ovn/commit/1c360bbd911cab9fadd6df8cd528d992ffa7a998

Fixes: 1c360bbd911cab9fadd6df8cd528d992ffa7a998 (Fix basic multicast flows for vxlan (non-vtep) tunnels)
Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>