Add support for port binding migration
When OVN moves a port binding to another node, options:requested-chassis
is updated to point to a new chassis, at which point northd translates
the change into SB updates; then the old chassis winds down flows
related to the binding, and the new chassis configures the binding in
the new location. While it works, it takes some time and is perceived as
temporary network downtime.

To reduce the downtime, this patch introduces a new option called
'migration-destination' that can be set on LS port, which makes the
second chassis to pre-configure the port binding in the new location
while the old binding still exists and serves the user. Then once
everything is ready on the other side, the user may move their payload
to the new location and, finally, wind down the original binding on the
first chassis.

We must guard the new binding location from prematurily communicating
with the outside world. This is achieved by installing drop flows into
tables 8 and 40 at the new location. These flows are removed when the
port at the new location issues a RARP packet, which indicates that the
payload has moved and is ready to serve from the new chassis. The packet
is caught by a separate controller() action flow. The action handler
then removes drop flows, as well as the rarp controller() action flow.
The handler also tags the binding with options:migration-unblocked=true.

Once this happens, the user may complete migration by unwinding the
binding at the original location. This is achieved by setting
options:requested-chassis to point to the new chassis, and unsetting
options:migration-destination.

Note: the design was inspired / re-invented based on previous
discussions of the use scenario, e.g. found at:

https://mail.openvswitch.org/pipermail/ovs-dev/2017-March/329865.html
https://etherpad.opendev.org/p/ovn_live_migration
https://bugzilla.redhat.com/show_bug.cgi?id=2012179

This is expected to be utilizied by OpenStack Neutron for VM live
migration.

TODO: perhaps postpone RARP to when drop flows are deleted from the
switch; then re-inject into pipeline. We can't just re-inject
the packet with a continuation pin inside the unblock_migration handler
because there's a lag between delete flow messages queued and applied by
the switch.

TODO: implement ddlog.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>