ovn-northd: Don't log transaction failures on standby instances.
We still need to try to ovsdb_idl_loop_commit_and_wait() on instances
that are in standby mode.  That's because they need to try to take the
lock.  But if they're in standby they won't actually build a transaction
json and will return early because they don't own the SB lock.

That's reported as an error by ovsdb_idl_loop_commit_and_wait() but we
shouldn't act on it.

Also, to ensure that no DB changes are missed, ovsdb_idl_track_clear()
should be called only on active instances.  The standby or paused ones
will get the incremental updates when they become active.  Otherwise we
would be forced to perform a full recompute when the instance becomes
active.

Fixes: 953832eb17f3 ("ovn-northd: Don't trigger recompute for transactions that are in progress.")
Fixes: 4597317f16d1 ("northd: Introduce incremental processing for northd")
Reported-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>