northd: Add feature to log reply and related ACL traffic.
It can be desirable for replies to stateful ACLs to be logged. And in
some cases, it can actually be a bit confusing why they aren't logged.
Consider a situation where a port group called "port_group" exists and
logical switch ports swp1 and swp2 belong to it. We create the following
ACL, where logging is enabled:

from-lport 100 'inport == @port_group' allow-stateless

swp1 sends traffic to swp2 and swp2 responds within the same connection.
In this case, the logs will show both the packets from swp1 to swp2, as
well as the response packets from swp2 to swp1.

Now change the ACL:

from-lport 100 'inport == @port_group' allow-related

Now with the same traffic pattern, the packets from swp1 to swp2 are
logged, but the packets from swp2 to swp1 are not. Why is that?

The reason is that as a shortcut, when stateful ACLs are used, a single
priority 65532 flow is programmed to allow reply traffic to pass. When
no stateful ACL is present, the reply traffic is at the mercy of the
stateless ACL on the reply. Therefore, with the stateful ACL, the reply
traffic is not actually hitting an ACL but is let through by default,
but with the stateless ACL, the reply traffic does hit the ACL
evaluation, so the reply traffic is logged.

This change adds a feature that allows for reply traffic to be
optionally logged for stateful ACLs, therefore allowing for the behavior
to be similar for both ACL types. Since logging reply traffic requires
adding more flows, it is not enabled by default. In order to have reply
traffic logged, the ACL must have logging enabled, be stateful, have a
label, and have the new log_related column set to true,

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>