ovn-controller: Delay ovn-installed if sb db is readonly
Before this patch, when all flows for a port have been installed in OVS,
ovn-installed was immediately written to the OVS DB; the notification of
this change to the OVN controller (handled by OVS interface handler)
might cause a full recompute if the SBDB is readonly / still handling
the chassis update in PortBinding.
With this patch, ovn-installed is not written to the OVS DB while the
SBDB is readonly.  This reduces the number of (un-necessary) full
recomputes.
Note that this was causing some spurious full recompute in test "ACL
with Port Group conjunction flow efficiency", causing the test to fail
from time to time. The test case was also updated as part of this patch
to check that ovn-installed was always properly set.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>