tests: log more info on OVN_CHECK_PACKETS* failure
Right now it logs diff between expected and observed only when enough
packets arrived. If less than expected packets arrived, only the
number of packets observed was logged, but not the packets themselves.
This patch makes the functions log the expected packets, the observed
packets, as well as the diff between them in this situation.

Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>