pinctrl: give a 'no such name' response when no DNS record is available
While the CentOS 8 requests an IPv4 DNS record, the response will be slow without a correct IPv6 DNS record on server. It will send a DNS request(type AAAA) message first but ovn controller will not reply. And the DNS request(type A) message will be sent until DNS request(type AAAA) is timeout.
So the DNS client may expect to receive a reponse with RCODE 3(RFC1035. Name Error) instead of nothing when DNS record is not available. And this patch is for that.

Signed-off-by: Yuan-96 <294754599@qq.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>