actions: Add action ct_lb_mark.
Add a new action ct_lb_mark, which is the same as ct_lb except that it
internally uses ct_mark to store the NAT flag, while ct_lb uses ct_label
for the same purpose. This will be used later to move the masked access
of ct_label to ct_mark while keeping the backward compatibility.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>