Added test cases with ovn-northd parallelization enabled
This will more or less double the number of test cases.
It is possible to select a reduce set of test cases using -k "keywords".
Keyword such as
        dp-groups=yes
        dp-groups=no
        parallelization=yes
        parallelization=no
        ovn-northd
        ovn-northd-ddlog
can be used to select a range of tests, as title is searched as well.

For instance, to run ovn-parallelization tests, with dp-groups enabled and ddlog disabled:
        make check TESTSUITEFLAGS="-k dp-groups=yes,parallelization=yes,\!ovn-northd-ddlog"

A few additional notes
- Do not use "ovn-northd" as a keyword, or the result might be different than expected.
  Even tough ovn-northd is added in all titles for the non ddlog tests, ovn-northd is also
  added in title of some other tests, for both ddlog and non ddlog (e.g. "check ovn-northd and ovn-controller version pinning").
- Running tests for both parallelization=yes and parallelization=no will not run all tests as
  some tests are not run through OVN_FOR_EACH_NORTHD. Instead use for instance
  make check  TESTSUITEFLAGS="-k parallelization=yes" and
  make check  TESTSUITEFLAGS="-k \!parallelization=yes" to run all tests

If less than four physical cores are available, then 1 socket/four cores dummy-numa
is used to excercice parallel northd

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>