northd: Don't set SB.Load_Balancer columns if not needed.
OVS commit 1cc618c32524 ("ovsdb-idl: Fix atomicity of writes that
don't change a column's value.") [0] explains why writes that don't
change a column's value cannot be optimized out early if the column is
read/write.

In northd, most tables have change tracking enabled, making all their
columns read/write.  That means that a write to a column (even if it
doesn't change the value) will add the row to the current transaction.
Validation is eventually performed before sending the transaction to the
server but if there are lots of such records this becomes costly.

Profiling what happens in northd when running with a NB database taken
from an ovn-k8s-like scale test (16K load balancers applied to 120
logical switches and routers) we notice that ovn-northd was always
writing to the SB.Load_Balancer columns even if nothing changed.

This commit changes that behavior and only writes to the
SB.Load_Balancer columns if needed.

Without this change, in our test server, with ovn-northd running
against NB database mentioned above, processing loop intervals were
~13 seconds.

With this change applied, loop intervals go down to ~7 seconds.

[0] https://github.com/openvswitch/ovs/commit/1cc618c32524

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>