ovn-nbctl: Add support for Port Mirroring
Changes done in ovn-nbctl to support creation  of mirror
and attach a source port for the same.

Example commands:
ovn-nbctl mirror-add mirror1 gre 0 to-lport sw0-port2
ovn-nbctl lsp-attach-mirror sw0-port1 mirror1

Changes are needed in ovn-sbctl and ovn-controller as well
to make it functional end to end. That needs to be done in
future commits as incremental changes.

Co-authored-by: Veda Barrenkala <vedabarrenkala@gmail.com>
Signed-off-by: Abhiram R N <abhiramrn@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>