ofctrl: Support ovn-ofctrl-wait-before-clear to avoid down time during upgrade.
Whenever OpenFlow connection between ovn-controller and OVS is
connected/reconnected, typically during ovn-controller/OVS
restart/upgrade, ovn-controller would:
1. clears all the existing flows after the initial hand-shaking
2. compute the new flows
3. install the new flows to OVS.

In large scale environments when there are a big number of flows
needs to be computed by ovn-controller, the step 2 above can take
very long time (from seconds to minutes, depending on the scale and
topology), which would cause a data plane down time.

The purpose of this patch is to avoid data plane down time during
restart/upgrade. It adds a new state S_WAIT_BEFORE_CLEAR to the ofctrl
state machine, so that ovn-controller waits for a period of time before
clearing the existing OVS flows while it is computing the new flows.
Ideally, ovn-controller should clear the flows after it completes the
new flows computing. However, it is difficult for ovn-controller to
judge if it has generated all the new flows (or the flows that are
sufficient to avoid down time), because flows computation depends on
iterations of SB monitor data processing and condition changes. So,
instead of try to determine by ovn-controller itself, this patch
provides a configuration "ovn-ofctrl-wait-before-clear" for users to
determine the feasible time to wait, according to the scale. As
mentioned in the document, the output of

  ovn-appctl -t ovn-controller stopwatch/show flow-generation

can help users determining what value to set.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>