ofctrl.c: Check installed flow when merging tracked flow changes.
During incremental flow processing, we track the OVS desired flow
changes so that we can incrementally install them to OVS. The function
merge_tracked_flows() is to merge the "delete and add/update" for the
same flows, to avoid unnecessary changes to OVS when flows are deleted
but added back in the same run. The function ensures that the flows
being deleted and added/updated are exactly the same, including actions.
It only compares the desired flows, assuming that the flow installed to
OVS is exactly the same as the desired flow being deleted.

However, this assumption can be wrong. If the same flow is already updated
before "delete and add/update" in the same run, the initial "update" is
not sent to OVS yet, but the change-tracking entry of that initial
"update" is already override by the "delete". This would result in lost
changes to OVS flows. This kind of problems can only be recovered by a
full recompute in ovn-controller.

This is more likely to cause real problems to branch-21.12 and earlier
releases, and less likely (if at all) to cause problems in later
releases, because since 22.03 we avoid repeated processing the same
logical flow in the same run, but it may still happen in corner cases
that hasn't been discovered.

This patch fixes the problem by adding the check for installed flows
before making the decision to merge the "delete and add/update" entries.
We will not merge if the installed flow doesn't match the desired flow
finally added/update.

Reported-by: François Rigault
Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2071272#c12
Reported-by: Numan Siddique <numans@ovn.org>
Reported-at: https://mail.openvswitch.org/pipermail/ovs-dev/2022-April/393256.html
Fixes: f4e508dd7 ("ofctrl.c: Merge opposite changes of tracked flows before installing.")
Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>